Setup initial test for entity movement
Created a simple test level scene, with an Entity node, that will move to a new random position, wait a few seconds with a timer, and repeat. I'm planning to use inheritance, and have the fake/imposter entity borrow the code from the main Entity class to do it's movement, but of course with added functionality. I understand the theory roughly, but I don't think I've ever really used inheritance on my own, so we'll see.