Continued work on initial test scene/prototype
- Moved imposter.gd script to the correct scripts folder
- Created 'imposter' version of the entity, which has a Visibility Notifier and signals to emit when the imposter is not in camera view
- Baked Occlusion for the basic level geometry
(Note: this is still buggy, as I noticed with some initial tests, mainly that it's too sensitive around edges, meaning that the Notifier will say it's visible in view when it isn't, it's still behind level geometry. I'm not sure if this is a quirk/quality setting issue with the Occulsion Culling system, or something with how the Visibility Notifer works with that system, it's not super egregious, but it is bad. The object is saying it's visible when it's certainly not, it's hidden behind a box object... I've looked at the settings and haven't found any easy way to fix besides manually tweaking each Occulder object to have a small AABB than it's actual geometry.
- Basic UI to change a Label on the screen to tell the player whether or not the Imposter entity is visible or not, using signals.
- Basic camera control that just rotates the camera around the center point of the world using A and D... Planning to implement a FPS Controller later